requirements.txt
# use pip show ... to check versions
alembic==1.18.1
arrow==1.3.0
binaryornot==0.4.4
blinker==1.9.0
certifi==2025.1.31
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cookiecutter==2.6.0
Flask==3.1.0
Flask-Login==0.6.3
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
markdown-it-py==3.0.0
markdown==3.7
MarkupSafe==3.0.2
mdurl==0.1.2
numpy==2.3.3
pandas==2.3.3
plotly==6.5.2
Pygments==2.19.1
python-dateutil==2.9.0.post0
python-slugify==8.0.4
PyYAML==6.0.2
requests==2.32.3
rich==14.0.0
Flask-Cors==4.0.0
six==1.17.0
SQLAlchemy==2.0.40
text-unidecode==1.3
types-python-dateutil==2.9.0.20241206
typing_extensions==4.13.0
urllib3==2.3.0
Werkzeug==3.1.3
